[FIX] donation: Fix Transifex
DB with demo data can be populated with current state, as there's
no chart template, and donation demo data search for a journal, which
is not present if no localization module is installed.

This trick assigns before that demo data a country to the main company
and execute again the autoinstall function through a helper method.