[FIX] donation by direct debit
For donation by debit order, we can't use twice the account "payment_debit_account_id", when
validating the donation AND when validation the debit order. We have to
use a different account. And, now that the OCA module
account_payment_order uses account.payment, this account must be a
receivable account.